# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Speedrulingo API", "version": "0.1.0"}, "paths": {"/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Health Health Get"}}}}}}}, "/ready": {"get": {"summary": "Ready", "operationId": "ready_ready_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"additionalProperties": {"type": "string"}, "type": "object", "title": "Response Ready Ready Get"}}}}}}}, "/api/v1/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "operationId": "register_api_v1_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_api_v1_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/auth/me": {"get": {"tags": ["auth"], "summary": "Me", "operationId": "me_api_v1_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MeResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/path": {"get": {"tags": ["content"], "summary": "Path", "operationId": "path_api_v1_path_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PathResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/course/current": {"get": {"tags": ["content"], "summary": "Current Course", "operationId": "current_course_api_v1_course_current_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CurrentCourseResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/units": {"get": {"tags": ["content"], "summary": "Units", "operationId": "units_api_v1_units_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/UnitSummary"}, "type": "array", "title": "Response Units Api V1 Units Get"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/units/{unit_id}": {"get": {"tags": ["content"], "summary": "Unit Detail", "operationId": "unit_detail_api_v1_units__unit_id__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "unit_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Unit Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UnitDetail"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/lessons/{lesson_id}/next-item": {"get": {"tags": ["learning"], "summary": "Next Item", "operationId": "next_item_api_v1_lessons__lesson_id__next_item_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "lesson_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Lesson Id"}}, {"name": "cursor", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 0, "default": 0, "title": "Cursor"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LessonItemResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/lessons/{lesson_id}/submit": {"post": {"tags": ["learning"], "summary": "Submit", "operationId": "submit_api_v1_lessons__lesson_id__submit_post", "security": [{"HTTPBearer": []}], "parameters": [{"name": "lesson_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Lesson Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SubmitRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SubmitResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/sentences/{sentence_id}/tokens": {"get": {"tags": ["explain"], "summary": "Sentence Tokens", "operationId": "sentence_tokens_api_v1_sentences__sentence_id__tokens_get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "sentence_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Sentence Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SentenceTokensResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/explain/token": {"post": {"tags": ["explain"], "summary": "Explain Token Route", "operationId": "explain_token_route_api_v1_explain_token_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExplainRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ExplainResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/kanji/lessons": {"get": {"tags": ["explain"], "summary": "Kanji Lessons", "operationId": "kanji_lessons_api_v1_kanji_lessons_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/KanjiLessonsResponse"}}}}}, "security": [{"HTTPBearer": []}]}}, "/api/v1/kanji/{kanji_char}": {"get": {"tags": ["explain"], "summary": "Kanji Detail", "operationId": "kanji_detail_api_v1_kanji__kanji_char__get", "security": [{"HTTPBearer": []}], "parameters": [{"name": "kanji_char", "in": "path", "required": true, "schema": {"type": "string", "title": "Kanji Char"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/KanjiDetailResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CurrentCourseResponse": {"properties": {"course_version_id": {"type": "string", "title": "Course Version Id"}, "course_code": {"type": "string", "title": "Course Code"}, "course_version": {"type": "integer", "title": "Course Version"}, "status": {"type": "string", "title": "Status"}, "current_unit_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Current Unit Id"}, "current_lesson_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Current Lesson Id"}}, "type": "object", "required": ["course_version_id", "course_code", "course_version", "status", "current_unit_id", "current_lesson_id"], "title": "CurrentCourseResponse"}, "ExplainRequest": {"properties": {"sentence_id": {"type": "string", "title": "Sentence Id"}, "token_surface": {"type": "string", "minLength": 1, "title": "Token Surface"}}, "additionalProperties": false, "type": "object", "required": ["sentence_id", "token_surface"], "title": "ExplainRequest"}, "ExplainResponse": {"properties": {"sentence_id": {"type": "string", "title": "Sentence Id"}, "sentence_ja": {"type": "string", "title": "Sentence Ja"}, "sentence_en": {"type": "string", "title": "Sentence En"}, "tokens": {"items": {"$ref": "#/components/schemas/SentenceTokenDTO"}, "type": "array", "title": "Tokens"}, "matching_tokens": {"items": {"$ref": "#/components/schemas/SentenceTokenDTO"}, "type": "array", "title": "Matching Tokens"}}, "type": "object", "required": ["sentence_id", "sentence_ja", "sentence_en", "tokens", "matching_tokens"], "title": "ExplainResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HintSpan": {"properties": {"token_start": {"type": "integer", "title": "Token Start"}, "token_end": {"type": "integer", "title": "Token End"}, "hint_text": {"type": "string", "title": "Hint Text"}, "hint_kind": {"type": "string", "title": "Hint Kind"}}, "type": "object", "required": ["token_start", "token_end", "hint_text", "hint_kind"], "title": "HintSpan"}, "ItemResult": {"properties": {"item_id": {"type": "string", "title": "Item Id"}, "expected_answer": {"type": "string", "title": "Expected Answer"}, "user_answer": {"type": "string", "title": "User Answer"}, "is_correct": {"type": "boolean", "title": "Is Correct"}}, "type": "object", "required": ["item_id", "expected_answer", "user_answer", "is_correct"], "title": "ItemResult"}, "KanjiDetailResponse": {"properties": {"kanji_char": {"type": "string", "title": "Kanji Char"}, "primary_meaning": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Primary Meaning"}, "usages": {"items": {"$ref": "#/components/schemas/KanjiUsageRow"}, "type": "array", "title": "Usages"}}, "type": "object", "required": ["kanji_char", "primary_meaning", "usages"], "title": "KanjiDetailResponse"}, "KanjiLessonSummary": {"properties": {"lesson_id": {"type": "string", "title": "Lesson Id"}, "unit_id": {"type": "string", "title": "Unit Id"}, "unit_order_index": {"type": "integer", "title": "Unit Order Index"}, "lesson_order_index": {"type": "integer", "title": "Lesson Order Index"}, "state": {"type": "string", "title": "State"}, "kanji_chars": {"items": {"type": "string"}, "type": "array", "title": "Kanji Chars"}}, "type": "object", "required": ["lesson_id", "unit_id", "unit_order_index", "lesson_order_index", "state", "kanji_chars"], "title": "KanjiLessonSummary"}, "KanjiLessonsResponse": {"properties": {"lessons": {"items": {"$ref": "#/components/schemas/KanjiLessonSummary"}, "type": "array", "title": "Lessons"}}, "type": "object", "required": ["lessons"], "title": "KanjiLessonsResponse"}, "KanjiUsageRow": {"properties": {"lesson_id": {"type": "string", "title": "Lesson Id"}, "unit_id": {"type": "string", "title": "Unit Id"}, "unit_order_index": {"type": "integer", "title": "Unit Order Index"}, "lesson_order_index": {"type": "integer", "title": "Lesson Order Index"}, "example_word_ja": {"type": "string", "title": "Example Word Ja"}, "example_reading": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Example Reading"}, "meaning_en": {"type": "string", "title": "Meaning En"}, "is_learned": {"type": "boolean", "title": "Is Learned"}}, "type": "object", "required": ["lesson_id", "unit_id", "unit_order_index", "lesson_order_index", "example_word_ja", "example_reading", "meaning_en", "is_learned"], "title": "KanjiUsageRow"}, "LessonItemResponse": {"properties": {"item_id": {"type": "string", "title": "Item Id"}, "lesson_id": {"type": "string", "title": "Lesson Id"}, "lesson_kind": {"type": "string", "title": "Lesson Kind"}, "item_type": {"type": "string", "enum": ["word_choice", "sentence_tiles", "kanji_kana_match"], "title": "Item Type"}, "order_index": {"type": "integer", "title": "Order Index"}, "cursor": {"type": "integer", "title": "Cursor"}, "total_items": {"type": "integer", "title": "Total Items"}, "is_last_item": {"type": "boolean", "title": "Is Last Item"}, "prompt_lang": {"type": "string", "title": "Prompt Lang"}, "answer_lang": {"type": "string", "title": "Answer Lang"}, "prompt_text": {"type": "string", "title": "Prompt Text"}, "sentence_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sentence Id"}, "sentence_ja_tokens": {"items": {"$ref": "#/components/schemas/SentenceTokenPreview"}, "type": "array", "title": "Sentence Ja Tokens"}, "sentence_ja_hints": {"items": {"$ref": "#/components/schemas/HintSpan"}, "type": "array", "title": "Sentence Ja Hints"}, "sentence_en_tokens": {"items": {"$ref": "#/components/schemas/SentenceTokenPreview"}, "type": "array", "title": "Sentence En Tokens"}, "answer_tiles": {"items": {"type": "string"}, "type": "array", "title": "Answer Tiles"}}, "type": "object", "required": ["item_id", "lesson_id", "lesson_kind", "item_type", "order_index", "cursor", "total_items", "is_last_item", "prompt_lang", "answer_lang", "prompt_text", "sentence_id", "sentence_ja_tokens", "sentence_ja_hints", "sentence_en_tokens", "answer_tiles"], "title": "LessonItemResponse"}, "LessonSummary": {"properties": {"id": {"type": "string", "title": "Id"}, "order_index": {"type": "integer", "title": "Order Index"}, "kind": {"type": "string", "title": "Kind"}, "state": {"type": "string", "title": "State"}, "is_locked": {"type": "boolean", "title": "Is Locked"}, "attempts_used": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Attempts Used"}}, "type": "object", "required": ["id", "order_index", "kind", "state", "is_locked"], "title": "LessonSummary"}, "LoginRequest": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "maxLength": 200, "minLength": 1, "title": "Password"}}, "additionalProperties": false, "type": "object", "required": ["email", "password"], "title": "LoginRequest"}, "MeResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "has_pro_sub": {"type": "boolean", "title": "Has Pro Sub"}, "is_admin": {"type": "boolean", "title": "Is Admin"}, "enrollment_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Enrollment Id"}, "course_version_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Course Version Id"}}, "type": "object", "required": ["id", "email", "has_pro_sub", "is_admin"], "title": "MeResponse"}, "PathResponse": {"properties": {"sections": {"items": {"$ref": "#/components/schemas/SectionUnits"}, "type": "array", "title": "Sections"}, "current_unit_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Current Unit Id"}, "current_lesson_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Current Lesson Id"}}, "type": "object", "required": ["sections", "current_unit_id", "current_lesson_id"], "title": "PathResponse"}, "RegisterRequest": {"properties": {"email": {"type": "string", "title": "Email"}, "password": {"type": "string", "maxLength": 200, "minLength": 8, "title": "Password"}}, "additionalProperties": false, "type": "object", "required": ["email", "password"], "title": "RegisterRequest"}, "SectionUnits": {"properties": {"id": {"type": "string", "title": "Id"}, "title": {"type": "string", "title": "Title"}, "units": {"items": {"$ref": "#/components/schemas/UnitSummary"}, "type": "array", "title": "Units"}}, "type": "object", "required": ["id", "title", "units"], "title": "SectionUnits"}, "SentencePreview": {"properties": {"id": {"type": "string", "title": "Id"}, "ja_text": {"type": "string", "title": "Ja Text"}, "en_text": {"type": "string", "title": "En Text"}}, "type": "object", "required": ["id", "ja_text", "en_text"], "title": "SentencePreview"}, "SentenceTokenDTO": {"properties": {"token_index": {"type": "integer", "title": "Token Index"}, "surface": {"type": "string", "title": "Surface"}, "lemma": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Lemma"}, "reading": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reading"}, "pos": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pos"}, "hints": {"items": {"type": "string"}, "type": "array", "title": "Hints"}}, "type": "object", "required": ["token_index", "surface", "lemma", "reading", "pos", "hints"], "title": "SentenceTokenDTO"}, "SentenceTokenPreview": {"properties": {"token_index": {"type": "integer", "title": "Token Index"}, "surface": {"type": "string", "title": "Surface"}, "lemma": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Lemma"}, "reading": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Reading"}, "pos": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Pos"}, "hints": {"items": {"type": "string"}, "type": "array", "title": "Hints", "default": []}}, "type": "object", "required": ["token_index", "surface", "lemma", "reading", "pos"], "title": "SentenceTokenPreview"}, "SentenceTokensResponse": {"properties": {"sentence_id": {"type": "string", "title": "Sentence Id"}, "sentence_ja": {"type": "string", "title": "Sentence Ja"}, "sentence_en": {"type": "string", "title": "Sentence En"}, "tokens": {"items": {"$ref": "#/components/schemas/SentenceTokenDTO"}, "type": "array", "title": "Tokens"}}, "type": "object", "required": ["sentence_id", "sentence_ja", "sentence_en", "tokens"], "title": "SentenceTokensResponse"}, "SubmitRequest": {"properties": {"answers": {"items": {"$ref": "#/components/schemas/SubmittedAnswer"}, "type": "array", "minItems": 1, "title": "Answers"}}, "additionalProperties": false, "type": "object", "required": ["answers"], "title": "SubmitRequest"}, "SubmitResponse": {"properties": {"lesson_id": {"type": "string", "title": "Lesson Id"}, "lesson_kind": {"type": "string", "title": "Lesson Kind"}, "score": {"type": "number", "title": "Score"}, "correct_items": {"type": "integer", "title": "Correct Items"}, "total_items": {"type": "integer", "title": "Total Items"}, "passed": {"type": "boolean", "title": "Passed"}, "progress_state": {"type": "string", "title": "Progress State"}, "item_results": {"items": {"$ref": "#/components/schemas/ItemResult"}, "type": "array", "title": "Item Results"}}, "type": "object", "required": ["lesson_id", "lesson_kind", "score", "correct_items", "total_items", "passed", "progress_state", "item_results"], "title": "SubmitResponse"}, "SubmittedAnswer": {"properties": {"item_id": {"type": "string", "title": "Item Id"}, "user_answer": {"type": "string", "minLength": 1, "title": "User Answer"}}, "additionalProperties": false, "type": "object", "required": ["item_id", "user_answer"], "title": "SubmittedAnswer"}, "TokenResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "const": "bearer", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "TokenResponse"}, "UnitDetail": {"properties": {"id": {"type": "string", "title": "Id"}, "section_id": {"type": "string", "title": "Section Id"}, "section_title": {"type": "string", "title": "Section Title"}, "unit_order_index": {"type": "integer", "title": "Unit Order Index"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "pattern_tags": {"items": {"type": "string"}, "type": "array", "title": "Pattern Tags"}, "theme_tags": {"items": {"type": "string"}, "type": "array", "title": "Theme Tags"}, "article_md": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Article Md"}, "lessons": {"items": {"$ref": "#/components/schemas/LessonSummary"}, "type": "array", "title": "Lessons"}, "sentence_samples": {"items": {"$ref": "#/components/schemas/SentencePreview"}, "type": "array", "title": "Sentence Samples"}}, "type": "object", "required": ["id", "section_id", "section_title", "unit_order_index", "title", "description", "pattern_tags", "theme_tags", "article_md", "lessons", "sentence_samples"], "title": "UnitDetail"}, "UnitSummary": {"properties": {"id": {"type": "string", "title": "Id"}, "section_id": {"type": "string", "title": "Section Id"}, "section_title": {"type": "string", "title": "Section Title"}, "unit_order_index": {"type": "integer", "title": "Unit Order Index"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "lesson_count": {"type": "integer", "title": "Lesson Count"}, "completed_lessons": {"type": "integer", "title": "Completed Lessons"}, "is_locked": {"type": "boolean", "title": "Is Locked"}, "is_completed": {"type": "boolean", "title": "Is Completed"}}, "type": "object", "required": ["id", "section_id", "section_title", "unit_order_index", "title", "description", "lesson_count", "completed_lessons", "is_locked", "is_completed"], "title": "UnitSummary"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"HTTPBearer": {"type": "http", "scheme": "bearer"}}}}